Turn ucare_cli into submodule
