+++ pyuploadcare/ucare_cli/__init__.py
@@ -5,26 +5,21 @@ import time
 import argparse
 import logging
 import os
-import sys
-import re
-import json
-from math import ceil
 
-import requests
 from six.moves import configparser
 from dateutil import parser
 
-from . import conf, __version__
-from .api import _build_user_agent
-from .api_resources import File, FileGroup, FileList, FilesStorage
-from .exceptions import UploadcareException, TimeoutError, UploadError
-
-
-def pprint(value):
-    print(json.dumps(value, indent=2))
+from pyuploadcare import conf, __version__
+from pyuploadcare.api_resources import File, FileGroup, FileList, FilesStorage
+from pyuploadcare.exceptions import (
+    UploadcareException, TimeoutError, UploadError
+)
+from pyuploadcare.ucare_cli.utils import pprint, bool_or_none, int_or_none
+from pyuploadcare.ucare_cli.sync import add_sync_files_parser
 
 
 logger = logging.getLogger('pyuploadcare')
+
 str_settings = (
     'pub_key',
     'secret',
@@ -36,14 +31,6 @@ bool_settings = (
     'verify_api_ssl',
     'verify_upload_ssl',
 )
-
-
-def bool_or_none(value):
-    return {'true': True, 'false': False}.get(value)
-
-
-def int_or_none(value):
-    return None if value.lower() == 'none' else int(value)
 
 
 def list_files(arg_namespace):
@@ -160,122 +147,6 @@ def create_group(arg_namespace):
     files = [File(uuid) for uuid in arg_namespace.paths]
     group = FileGroup.create(files)
     pprint(group.info())
-
-
-def sync_files(arg_namespace):
-    if arg_namespace.uuids:
-        files = (File(uuid) for uuid in arg_namespace.uuids)
-    else:
-        files = FileList()
-
-    session = requests.Session()
-    session.headers.update({'User-Agent': _build_user_agent()})
-
-    def _get(url, max_retry=3):
-        response = None
-
-        for i in range(max_retry):
-            try:
-                response = session.get(url, stream=True,
-                                       verify=conf.verify_api_ssl)
-            except requests.exceptions.ConnectionError as e:
-                pprint('Connection Error: {0}'.format(e))
-                pprint('Retry..')
-                time.sleep(i ** 2)
-                continue
-            else:
-                break
-
-        if response is None:
-            pprint('Can\'t fetch URL: {0}'.format(url))
-            pprint('Skip it.')
-            return
-
-        try:
-            response.raise_for_status()
-        except requests.HTTPError as e:
-            pprint(('Can\'t download file: `{0}`. '
-                    'Origin error: {1}').format(url, e))
-            return
-
-        return response
-
-    for f in files:
-        if f.is_image() and arg_namespace.effects:
-            f.default_effects = arg_namespace.effects.lstrip('-/')
-
-        local_filepath = build_filepath(arg_namespace.path, f)
-        dirname = os.path.dirname(local_filepath)
-
-        if dirname and not os.path.exists(dirname):
-            os.makedirs(dirname)
-
-        if os.path.exists(local_filepath) and not arg_namespace.replace:
-            pprint(
-                'File `{0}` already exists. '
-                'To override it use `--replace` option'.format(
-                    local_filepath))
-            continue
-
-        url = f.cdn_url
-        response = _get(url)
-
-        if response:
-            save_file_locally(local_filepath, response, f.size())
-
-
-PATTERNS_REGEX = re.compile(r'(\${\w+})')
-PATTERNS_MAPPING = {
-    '${uuid}': lambda f: f.uuid,
-    '${filename}': lambda f: f.filename(),
-    '${effects}': lambda f: f.default_effects,
-    '${ext}': lambda f: os.path.splitext(f.filename() or '')[-1],
-}
-DEFAULT_PATTERN_FILENAME = '${uuid}${ext}'
-
-
-def build_filepath(path, file_):
-    if not PATTERNS_REGEX.findall(path):
-        path = os.path.join(path, DEFAULT_PATTERN_FILENAME)
-
-    def _replace(mobj):
-        pattern_name = mobj.group(0)
-        if pattern_name in PATTERNS_MAPPING:
-            pattern = PATTERNS_MAPPING[pattern_name]
-            return pattern(file_)
-        return pattern_name
-
-    return os.path.normpath(PATTERNS_REGEX.sub(_replace, path))
-
-
-def save_file_locally(fname, response, size):
-    chunk_size = 1024
-    with open(fname, 'wb') as lf:
-        for chunk in bar(response.iter_content(chunk_size),
-                         ceil(size / float(chunk_size)),
-                         fname):
-            lf.write(chunk)
-
-
-def bar(iter_content, parts, title=''):
-    parts = max(float(parts), 1.0)
-    cells = 10
-    progress = 0
-    step = cells / parts
-
-    draw = lambda progress: sys.stdout.write(
-        '\r[{0:10}] {1:.2f}% {2}'.format(
-            '#'*int(progress), progress * cells, title))
-
-    for chunk in iter_content:
-        yield chunk
-
-        progress += step
-        draw(progress)
-        sys.stdout.flush()
-
-    draw(cells)
-    print('')
 
 
 def ucare_argparser():
@@ -419,25 +290,7 @@ def ucare_argparser():
     subparser.add_argument('paths', nargs='+', help='file paths')
 
     # Sync files
-    subparser = subparsers.add_parser('sync', help='sync files')
-    subparser.set_defaults(func=sync_files)
-    subparser.add_argument('path', nargs='?', help=(
-        'Local path. It can contains special patterns like: {0} '
-        'Default is {1}'.format(
-            ' '.join(PATTERNS_MAPPING.keys()),
-            DEFAULT_PATTERN_FILENAME)
-    ), default='.')
-    subparser.add_argument('--replace', help='replace exists files',
-                           default=False, action='store_true')
-    subparser.add_argument('--uuids', nargs='+',
-                           help='list of file\'s uuids for sync',)
-    subparser.add_argument('--effects', help=(
-        'apply effects for synced images.'
-        'Note that effects will apply to images only.'
-        'For more information look at: '
-        'https://uploadcare.com/documentation/cdn/  '
-        'Example: --effects=resize/200x/-/rotate/90/'
-    ))
+    add_sync_files_parser(subparsers)
 
     # common arguments
     parser.add_argument(
